Fallback search mechanism for DOIs not on crossref
While testing DOIsh against a dataset I noticed that, while DOIsh uses the
commonly accepted as being the most comprehensive DOI search interface
(crossref), there is a non-neglegible amount of papers that still do not appear
on crossref's results. Reading some things on the web on how academic
researchers (I'm not one) are meant to deal with this cases, the answer is
always consistent, even if disturbing: search the DOI on Google. And, true
enough, for the few cases I've tried it, it was an effective method of reaching
the data, even if not efficient.

This commit adds Google Search fallback mechanism. It's a start, even a lot of
work can be done to enhance this: in particular, trying to fetch data on who's
the publisher.